Split MacroDisplay code and create MacroGraph component
