+++ src/components/MacroGraph/index.ts
@@ -1,0 +1,1 @@
+export { default as MacroGraph } from './MacroGraph';
